requirements.txt
# Core ML - Essential only
pandas==2.1.4
numpy>=1.25.0,<3.0.0
scikit-learn==1.3.2
lightgbm==4.1.0
optuna==3.5.0
shap==0.43.0

# Web & API - Ultra-lightweight
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# LLM Integration - Single provider
groq==0.4.1

# Utilities - Minimal set
python-dotenv==1.0.0
joblib==1.3.2
plotly==5.17.0
protobuf>=5.0,<7.0
